requirements.txt
# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Data handling
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image processing
opencv-python>=4.8.0
numpy>=1.26.4  # Updated to match instanseg-torch requirements
Pillow>=10.0.0
openslide-python>=1.3.0
scikit-image>=0.22.0  # Updated to match instanseg-torch requirements

# InstanSeg for cell segmentation
instanseg-torch
tiffslide  # For WSI image reading support

# Task queue and async
redis>=5.0.0
celery>=5.3.0
# asyncio is built-in, no need to install

# Database (optional, can use in-memory for MVP)
sqlalchemy>=2.0.0
alembic>=1.12.0

# Monitoring and metrics (bonus)
prometheus-client>=0.19.0

# Utilities
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

